feat(sdk): promote Tooltip to public SDK surface (0.7.0)
Move the host's Tooltip component to @silo-code/sdk so silo.* and
third-party extensions can use the same styled hover popup without
reaching into @silo-code/extension-host/internal.

- packages/sdk: add Tooltip.tsx (component logic, no CSS import);
  export from barrel; add react-dom peer dep; bump to 0.7.0
- extension-host/Tooltip.tsx: thin re-export from SDK that keeps the
  CSS load (host still owns the stylesheet)
- sdk-internal.ts: update comment — Tooltip is now public, the
  internal re-export is convenience for core.* extensions
- file-explorer/TreeNodes.tsx: replace native title= with <Tooltip>
  on the four root-action buttons (New File / New Folder / Refresh /
  Collapse All) — tested in the running app
- docs: add api/other/tooltip.md; link from overview table, sidebar,
  and roadmap; regenerate TypeDoc type reference
