+++ apps/docs/.vitepress/config.ts
@@ -54,6 +54,7 @@ const apiSidebar = [
       { text: "ctx.getExtension", link: "/api/other/get-extension" },
       { text: "useServiceState", link: "/api/other/use-service-state" },
       { text: "useFocusGroup", link: "/api/other/use-focus-group" },
+      { text: "Tooltip", link: "/api/other/tooltip" },
       {
         text: "focusGroupNextIndex",
         link: "/api/other/focus-group-next-index",
