+++ apps/docs/api/types/interfaces/ShowMenuOptions.md
@@ -59,7 +59,7 @@ Anchor element to hang the menu off (for a button dropdown).
 ### align?
 
 ```ts
-optional align?: "start" | "end";
+optional align?: "end" | "start";
 ```
 
 Defined in: [packages/sdk/src/ui-service.ts:253](https://github.com/silo-code/silo/blob/main/packages/sdk/src/ui-service.ts#L253)
